requirements.txt
fire
llama-cpp-python
openai-whisper
pyaudio
pyyaml
SpeechRecognition
torch
